refactor: migrate footer styles from CSS to SCSS for improved maintainability
