+++ src/entities/website/Footer/Navigation/FooterNavigation.tsx
@@ -2,7 +2,7 @@ import { Flex, Group } from '@mantine/core';
 
 import FooterNavigationSection from '@/shared/components/FooterNavigationSection';
 import { websiteSections } from '@/shared/constants/urls';
-import styles from '@/entities/website/Footer/styles/website-footer.module.css';
+import styles from '@/entities/website/Footer/styles/website-footer.module.scss';
 
 export default function FooterNavigation() {
     return (
