--- src/entities/website/Footer/styles/website-footer.module.css
@@ -1,116 +1,0 @@
-.footerFeedBack,
-.footerContainer {
-    border-radius: 32px;
-    padding: 24px 19px;
-    background-color: var(--mantine-color-dark-8);
-    color: var(--mantine-color-gray-0);
-}
-
-.footerFeedBack {
-    padding: 12px 8px 8px;
-}
-
-.footerFeedBackLabel {
-    text-align: center;
-}
-
-.footerNavigationItems {
-    flex-direction: column;
-    gap: 1rem;
-}
-
-.footerNavigationItem {
-    text-decoration: none;
-}
-
-.footerContacts {
-    gap: 2rem;
-    flex-direction: column;
-}
-
-.footerLogo {
-    width: 72px;
-    height: 72px;
-}
-
-.footerSocialMedias {
-    gap: 2rem;
-}
-
-/* Tablet only */
-@media (min-width: $mantine-breakpoint-sm) and (max-width: $mantine-breakpoint-lg) {
-    .footerContacts {
-        gap: 0;
-        flex-direction: row;
-        justify-content: space-between;
-        width: 100%;
-    }
-
-    .footerNavigationContainer {
-        justify-content: center;
-        gap: 1rem;
-    }
-
-    .footerNavigationItems {
-        flex-direction: row;
-    }
-
-    .footerContainer {
-        gap: 2.5rem;
-    }
-}
-
-@media (min-width: $mantine-breakpoint-lg) {
-    .footerContacts {
-        gap: 1rem;
-        flex-direction: column;
-    }
-
-    .footerContainer {
-        padding: 30px 100px;
-        gap: 2rem;
-    }
-
-    .footerNavigationItems {
-        flex-direction: column;
-    }
-
-    .footerFeedBack {
-        padding: 0;
-        width: 250px;
-    }
-
-    .footerFeedBackLabel {
-        text-align: start;
-    }
-
-    .footerLogo {
-        width: 250px;
-        height: 250px;
-    }
-
-    .footerSocialMedias {
-        justify-content: space-between;
-        width: 100%;
-    }
-}
-
-@media (min-width: $mantine-breakpoint-xl) {
-    .footerFeedBack {
-        width: 350px;
-    }
-}
-
-@media (min-width: $mantine-breakpoint-xxl) {
-    .footerNavigationContainer {
-        flex-direction: column;
-    }
-
-    .footerNavigationItems {
-        flex-direction: row;
-    }
-
-    .footerFeedBack {
-        width: 450px;
-    }
-}
