+++ src/shared/styles/_variables.scss
@@ -1,0 +1,21 @@
+// Mantine breakpoints
+$mantine-breakpoint-xs: 576px;
+$mantine-breakpoint-sm: 768px;
+$mantine-breakpoint-md: 992px;
+$mantine-breakpoint-lg: 1200px;
+$mantine-breakpoint-xl: 1440px;
+$mantine-breakpoint-xxl: 1920px;
+
+// Mantine spacing
+$mantine-spacing-xs: 0.625rem;
+$mantine-spacing-sm: 0.75rem;
+$mantine-spacing-md: 1rem;
+$mantine-spacing-lg: 1.25rem;
+$mantine-spacing-xl: 2rem;
+
+// Mantine radius
+$mantine-radius-xs: 0.125rem;
+$mantine-radius-sm: 0.25rem;
+$mantine-radius-md: 0.5rem;
+$mantine-radius-lg: 1rem;
+$mantine-radius-xl: 2rem;
